+++ resharper/resharper-unity/src/AsmDefCommon/Psi/DeclaredElements/IAsmDefDeclaredElement.cs
@@ -1,0 +1,10 @@
+using JetBrains.ReSharper.Psi;
+
+namespace JetBrains.ReSharper.Plugins.Unity.AsmDefCommon.Psi.DeclaredElements
+{
+    public interface IAsmDefDeclaredElement : IDeclaredElement
+    {
+        IPsiSourceFile SourceFile { get; }
+        int DeclarationOffset { get; }
+    }
+}