+++ resharper/resharper-unity/src/AsmDefNew/Feature/Services/ContextSearch/AsmDefImplementationContextSearch.cs
@@ -1,0 +1,12 @@
+using JetBrains.Application;
+using JetBrains.ReSharper.Plugins.Unity.AsmDefCommon.Feature.Services.ContextSearch;
+using JetBrains.ReSharper.Plugins.Unity.JsonNew.Psi;
+
+namespace JetBrains.ReSharper.Plugins.Unity.AsmDefNew.Feature.Services.ContextSearch
+{
+    [ShellFeaturePart]
+    public class AsmDefImplementationContextSearch: AsmDefImplementationContextSearchBase<JsonNewLanguage>
+    {
+        
+    }
+}