+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/ideaInterop/fileTypes/asmdef/AsmDefLanguage.kt
@@ -1,6 +1,0 @@
-package com.jetbrains.rider.plugins.unity.ideaInterop.fileTypes.asmdef
-
-import com.jetbrains.rider.ideaInterop.fileTypes.RiderLanguageBase
-
-// Use the JSON language on the backend
-object AsmDefLanguage : RiderLanguageBase("AsmDef", "JSON")