+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/ideaInterop/fileTypes/asmdef/AsmDefSyntaxHighlighter.kt
@@ -1,5 +1,0 @@
-package com.jetbrains.rider.plugins.unity.ideaInterop.fileTypes.asmdef
-
-import com.jetbrains.rider.ideaInterop.fileTypes.RiderDummySyntaxHighlighter
-
-class AsmDefSyntaxHighlighter : RiderDummySyntaxHighlighter(AsmDefLanguage)