+++ rider/src/main/kotlin/com/jetbrains/rider/settings/templates/UnityTemplatesOptionPage.kt
@@ -1,0 +1,16 @@
+package com.jetbrains.rider.settings.templates
+
+import com.intellij.openapi.options.Configurable
+import com.jetbrains.rider.settings.simple.SimpleOptionsPage
+
+class UnityFileTemplatesOptionPage: SimpleOptionsPage("Unity", "RiderUnityFileTemplatesSettings"), Configurable.NoScroll {
+    override fun getId(): String {
+        return "RiderUnityFileTemplatesSettings"
+    }
+}
+
+class UnityLiveTemplatesOptionPage: SimpleOptionsPage("Unity", "RiderUnityLiveTemplatesSettings"), Configurable.NoScroll {
+    override fun getId(): String {
+        return "RiderUnityLiveTemplatesSettings"
+    }
+}